Fixed tmpRatio formula
tmpRatio nows gives a penalty on the characters alike between the query and the entry divided by the size of the entry.

This was changed from the old formula which gave a penalty based upon the size of the query divided by the size of the entry. This would skew results towards smaller entry sizes, if the query was larger than the entry.