Erryial added in support for a more scaleable code.
I finally figured out how to create a vector of vector that supports different platforms like game and book inside of 1 vector. This was originally an issue because vector don't store multiple different types like int and string, and in our case, books and games.

The solution was a tad bit complex and probably a bit unethical but it seems to working as of this minute in production.